TagEmit strongly typed; interface cleanup
