+++ src/app/components/sheet/sheet.component.ts
@@ -7,15 +7,10 @@ import { ManualTagsService } from '../tags-manual/manual-tags.service';
 import { FilePathService } from '../views/file-path.service';
 
 import { StarRating, ImageElement } from '../../../../interfaces/final-object.interface';
-
-import { metaAppear, textAppear, modalAnimation } from '../../common/animations';
+import { TagEmit, TagEmission } from '../../../../interfaces/shared-interfaces';
 import { YearEmission } from '../views/details/details.component';
 
-export interface TagEmission {
-  index: number;
-  tag: string;
-  type: 'add' | 'remove';
-}
+import { metaAppear, textAppear, modalAnimation } from '../../common/animations';
 
 export interface StarEmission {
   index: number;
@@ -46,7 +41,7 @@ export class SheetComponent implements OnInit {
   @Output() editFinalArrayStars = new EventEmitter<StarEmission>();
   @Output() editFinalArrayTag = new EventEmitter<TagEmission>();
   @Output() editFinalArrayYear = new EventEmitter<YearEmission>();
-  @Output() filterTag = new EventEmitter<object>();
+  @Output() filterTag = new EventEmitter<TagEmit>();
   @Output() openVideoAtTime = new EventEmitter<object>();
 
   @Input() video: ImageElement;
