+++ src/app/components/tags-auto/tag-display.pipe.ts
@@ -2,15 +2,9 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 import { autoFileTagsRegex } from './autotags.service';
 
-import { ImageElement } from '../../../../interfaces/final-object.interface';
-
 import { Colors } from '../../common/colors';
-
-export interface Tag {
-  name: string;
-  colour: string;
-  removable: boolean;
-}
+import { ImageElement } from '../../../../interfaces/final-object.interface';
+import { Tag } from '../../../../interfaces/shared-interfaces';
 
 @Pipe({
   name: 'tagDisplayPipe'
